fin de test AccessSiteDao

+++ src/data/AccessSite.java
@@ -30,7 +30,7 @@ public class AccessSite {
         return user_id;
     }
 
-    public void setUser_id(int user_id) {
+    public void setUser_id(long user_id) {
         this.user_id = user_id;
     }
